+++ core/watcher.ts
@@ -159,7 +159,7 @@ export class SiteWatcher implements Watcher {
     await this.dispatchEvent({ type: "start" });
     this.site.addEventListener("afterUpdate", (event) => {
       const files = new Set([
-        ...event.pages!.map((page) => page.outputPath),
+        ...event.pages!.map((page) => page.outputPath!),
         ...event.staticFiles!.map((file) => file.outputPath!),
       ]);
       this.dispatchEvent({ type: "change", files });