+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [1.13.1] - Unreleased
+### Removed
+- `Page.isHtml` property added in v1.13.0.
+
 ### Fixed
 - Updated the dependencies: `std`, `minify_html`, `sheetjs`.
 - Updated `highlight.js` types.